Moves Filter Service to Services Directory
