+++ src/lib/components/market/item_row_wrapper.ts
@@ -9,7 +9,7 @@ import {gFloatFetcher} from '../../float_fetcher/float_fetcher';
 import {ItemInfo} from '../../bridge/handlers/fetch_inspect_info';
 import {getMarketInspectLink, inlineEasyInspect, inlineStickers} from './helpers';
 import {formatSeed, renderClickableRank} from '../../utils/skin';
-import {gFilterService} from '../../filter/service';
+import {gFilterService} from '../../services/filter';
 import {AppId, ContextId, Currency} from '../../types/steam_constants';
 import {defined} from '../../utils/checkers';
